Let the transient heap belong to objspace
